requirements.txt
# Flask Backend
Flask==3.0.0
Flask-CORS==4.0.0

# Database
pandas==2.1.4

# Machine Learning
tensorflow==2.18.0
numpy==1.26.2
scikit-learn==1.3.2

# API & HTTP
requests==2.31.0

# Background Tasks
APScheduler==3.10.4

# Gemini AI
google-generativeai>=0.3.0

# Environment Variables
python-dotenv==1.0.0

# Utilities
pytz==2023.3